+++ src/Datasets/Labeled.php
@@ -274,7 +274,7 @@ class Labeled extends DataFrame implements Dataset
                 . " cannot be less than 1, $n given.");
         }
 
-        return $this->slice($n, $this->numRows());
+        return $this->slice(-$n, $this->numRows());
     }
 
     /**
@@ -745,6 +745,37 @@ class Labeled extends DataFrame implements Dataset
     }
 
     /**
+     * Generate a random subset without replacement.
+     *
+     * @param int $n
+     * @throws \InvalidArgumentException
+     * @return self
+     */
+    public function randomSubsetWithoutReplacement(int $n) : self
+    {
+        if ($n < 1) {
+            throw new InvalidArgumentException('Cannot generate subset'
+                . " of less than 1 sample, $n given.");
+        }
+
+        if ($n > $this->numRows()) {
+            throw new InvalidArgumentException('Cannot generate subset'
+                . " of more than {$this->numRows()}, $n given.");
+        }
+
+        $indices = array_rand($this->samples, $n);
+
+        $indices = is_array($indices) ? $indices : [$indices];
+
+        $indices = array_flip($indices);
+
+        return self::quick(
+            array_intersect_key($this->samples, $indices),
+            array_intersect_key($this->labels, $indices)
+        );
+    }
+
+    /**
      * Generate a random subset with replacement.
      *
      * @param int $n
@@ -754,7 +785,7 @@ class Labeled extends DataFrame implements Dataset
     public function randomSubsetWithReplacement(int $n) : self
     {
         if ($n < 1) {
-            throw new InvalidArgumentException('Cannot generate a'
+            throw new InvalidArgumentException('Cannot generate'
                 . " subset of less than 1 sample, $n given.");
         }
 
@@ -783,7 +814,7 @@ class Labeled extends DataFrame implements Dataset
     public function randomWeightedSubsetWithReplacement(int $n, array $weights) : self
     {
         if ($n < 1) {
-            throw new InvalidArgumentException('Cannot generate a'
+            throw new InvalidArgumentException('Cannot generate'
                 . " subset of less than 1 sample, $n given.");
         }